Rename `file_operations` to `io_operations`
