+++ crates/forge/tests/e2e/mod.rs
@@ -1,5 +1,5 @@
 mod collection;
 pub(crate) mod common;
-mod file_operations;
+mod io_operations;
 mod running;
 mod workspaces;
